fix(common): handle history `pushState|replaceState` errors in restricted environments
This commit wraps `history.pushState` and `history.replaceState` in try-catch blocks and rethrows an Angular runtime error with a meaningful error code.

In Firefox, the internal check looks like this:

```cpp
if (!win->HasActiveDocument()) {
  aRv.Throw(NS_ERROR_DOM_SECURITY_ERR);
  return;
}
```

This results in a generic "Operation is insecure" error, which may be unhelpful or misleading in error aggregators like Rollbar.

Unfortunately, it's not possible to perform reliable end-to-end testing of this fix.